# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations/{organizationId}/sites": {
      "post": {
        "description": "Requiere el permiso `site.create`. El slug es la identidad pública del sitio (`slug.dominio`) y es único en toda la plataforma, no solo dentro de la organización.",
        "operationId": "SitesController_create",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Organización dueña del recurso. Resuelve el tenant antes de tocar nada (ADR-002).",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": false,
            "description": "Obligatoria con valor `impulza-one` en toda petición que modifica estado. Ver CsrfGuard.",
            "schema": {
              "type": "string",
              "enum": [
                "impulza-one"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "slug": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 63,
                    "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
                  }
                },
                "required": [
                  "name",
                  "slug"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sitio creado en estado `DRAFT`.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "organizationId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "DRAFT",
                        "PUBLISHED",
                        "ARCHIVED"
                      ]
                    },
                    "themeId": {
                      "nullable": true,
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    }
                  },
                  "required": [
                    "id",
                    "organizationId",
                    "name",
                    "slug",
                    "status",
                    "themeId",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (detalle en `issues`)."
          },
          "401": {
            "description": "Sin sesión válida."
          },
          "403": {
            "description": "Sin membresía activa en la organización, sin el permiso requerido, o petición sin la cabecera anti-CSRF."
          },
          "409": {
            "description": "Ese slug ya lo usa otro sitio, o lo ocupa una redirección viva de otro sitio."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Crear un sitio",
        "tags": [
          "sites"
        ]
      },
      "get": {
        "description": "Solo pide membresía activa: ANALYST y SUPPORT necesitan verlos para hacer su trabajo.",
        "operationId": "SitesController_list",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Organización dueña del recurso. Resuelve el tenant antes de tocar nada (ADR-002).",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": false,
            "description": "Obligatoria con valor `impulza-one` en toda petición que modifica estado. Ver CsrfGuard.",
            "schema": {
              "type": "string",
              "enum": [
                "impulza-one"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sitios de la organización, archivados incluidos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "organizationId": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "name": {
                        "type": "string"
                      },
                      "slug": {
                        "type": "string"
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "DRAFT",
                          "PUBLISHED",
                          "ARCHIVED"
                        ]
                      },
                      "themeId": {
                        "nullable": true,
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      }
                    },
                    "required": [
                      "id",
                      "organizationId",
                      "name",
                      "slug",
                      "status",
                      "themeId",
                      "createdAt",
                      "updatedAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (detalle en `issues`)."
          },
          "401": {
            "description": "Sin sesión válida."
          },
          "403": {
            "description": "Sin membresía activa en la organización, sin el permiso requerido, o petición sin la cabecera anti-CSRF."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Listar los sitios de la organización",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/v1/organizations/{organizationId}/sites/{siteId}": {
      "get": {
        "operationId": "SitesController_getOne",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Organización dueña del recurso. Resuelve el tenant antes de tocar nada (ADR-002).",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "siteId",
            "required": true,
            "in": "path",
            "description": "Sitio de la organización.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": false,
            "description": "Obligatoria con valor `impulza-one` en toda petición que modifica estado. Ver CsrfGuard.",
            "schema": {
              "type": "string",
              "enum": [
                "impulza-one"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "El sitio solicitado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "organizationId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "DRAFT",
                        "PUBLISHED",
                        "ARCHIVED"
                      ]
                    },
                    "themeId": {
                      "nullable": true,
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    }
                  },
                  "required": [
                    "id",
                    "organizationId",
                    "name",
                    "slug",
                    "status",
                    "themeId",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (detalle en `issues`)."
          },
          "401": {
            "description": "Sin sesión válida."
          },
          "403": {
            "description": "Sin membresía activa en la organización, sin el permiso requerido, o petición sin la cabecera anti-CSRF."
          },
          "404": {
            "description": "Sitio no encontrado, o pertenece a otra organización (ADR-002)."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Leer un sitio",
        "tags": [
          "sites"
        ]
      },
      "patch": {
        "description": "Requiere el permiso `site.update`. Cambiarle el slug a un sitio **publicado** deja una redirección desde el anterior, para no romper en silencio los enlaces ya compartidos; un sitio en borrador nunca fue alcanzable, así que no genera redirección. Un cuerpo vacío es 400, no un 200 que no hace nada.",
        "operationId": "SitesController_update",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Organización dueña del recurso. Resuelve el tenant antes de tocar nada (ADR-002).",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "siteId",
            "required": true,
            "in": "path",
            "description": "Sitio de la organización.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": false,
            "description": "Obligatoria con valor `impulza-one` en toda petición que modifica estado. Ver CsrfGuard.",
            "schema": {
              "type": "string",
              "enum": [
                "impulza-one"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "slug": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 63,
                    "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sitio actualizado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "organizationId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "DRAFT",
                        "PUBLISHED",
                        "ARCHIVED"
                      ]
                    },
                    "themeId": {
                      "nullable": true,
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    }
                  },
                  "required": [
                    "id",
                    "organizationId",
                    "name",
                    "slug",
                    "status",
                    "themeId",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (detalle en `issues`)."
          },
          "401": {
            "description": "Sin sesión válida."
          },
          "403": {
            "description": "Sin membresía activa en la organización, sin el permiso requerido, o petición sin la cabecera anti-CSRF."
          },
          "404": {
            "description": "Sitio no encontrado, o pertenece a otra organización (ADR-002)."
          },
          "409": {
            "description": "Ese slug ya lo usa otro sitio, o lo ocupa una redirección viva de otro sitio."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Editar un sitio",
        "tags": [
          "sites"
        ]
      }
    },
...
  },
  "info": {
    "title": "Impulza One API",
    "description": "API REST de Impulza One. Todo cuelga de `/api/v1` salvo `/health`, que es un endpoint de\ninfraestructura sin sesión ni CSRF.\n\n**Autenticación**: cookie de sesión `HttpOnly` (`impulza_session`), emitida por\n`POST /auth/login`. No hay tokens en el cuerpo ni en cabeceras de autorización.\n\n**CSRF**: toda petición que modifica estado (POST, PATCH, PUT, DELETE) debe enviar la\ncabecera `X-Requested-With: impulza-one`. Un formulario o una imagen entre sitios no puede\nfijar cabeceras propias, y una petición `fetch` con cabecera custom dispara preflight CORS,\nbloqueado salvo origen autorizado. Sin ella la respuesta es 403.\n\n**Multi-tenant**: los recursos de negocio cuelgan de `/organizations/:organizationId/…`.\nUn recurso de otra organización responde **404 y no 403**, para no confirmar que existe\n(ADR-002).\n\n**Errores de validación**: 400 cuando el cuerpo no cumple el esquema de entrada, 422 cuando\ncumple la forma pero viola una regla de dominio (contraste insuficiente en un tema,\nconfiguración inválida para un tipo de bloque). Ambos traen `issues` con el campo exacto.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "sesion": {
        "type": "apiKey",
        "in": "cookie",
        "name": "impulza_session"
      }
    },
...
  }
}